+++ features/transaction/purchase_platform/src/main/java/com/tokopedia/purchase_platform/features/one_click_checkout/order/data/CartDataResponse.kt
@@ -1,15 +1,13 @@
 package com.tokopedia.purchase_platform.features.one_click_checkout.order.data
 
 import com.google.gson.annotations.SerializedName
-import com.tokopedia.purchase_platform.features.checkout.data.model.response.shipment_address_form.Shop
-import com.tokopedia.purchase_platform.features.checkout.data.model.response.shipment_address_form.ShopShipment
 import com.tokopedia.purchase_platform.features.checkout.data.model.response.shipment_address_form.Warehouse
 
 data class CartDataResponse(
         @SerializedName("errors")
         val errors: List<String> = emptyList(),
         @SerializedName("cart_id")
-        val cartId: Int = 0,
+        val cartId: Long = 0,
         @SerializedName("product")
         val product: ProductDataResponse = ProductDataResponse(),
         @SerializedName("shop")
