+++ features/transaction/purchase_platform/src/main/java/com/tokopedia/purchase_platform/features/one_click_checkout/order/view/card/OrderProductCard.kt
@@ -194,7 +194,7 @@ class OrderProductCard(private val view: View, private val listener: OrderProduc
     }
 
     private fun showPrice() {
-        var productPrice = product.productPrice.toLong()
+        var productPrice = product.productPrice
         if (product.wholesalePrice.isNotEmpty()) {
             for (wholesalePrice in product.wholesalePrice) {
                 if (product.quantity.orderQuantity >= wholesalePrice.qtyMin) {
